requirements.txt
# Core Framework
streamlit>=1.32.0

# Google Gemini AI
google-generativeai>=0.5.0

# OpenAI (for embeddings)
openai>=1.0.0

# Vector Database
chromadb>=0.4.0

# OCR
easyocr>=1.7.0

pydub>=0.25.0
soundfile>=0.12.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Symbolic Math
sympy>=1.12

# Environment Management
python-dotenv>=1.0.0

# Additional utilities
requests>=2.31.0
